Add Python 3.5/3.6 support
Add Python 3.5/3.6 support to the python-3 front end.  As far as I
can tell, there are two missing syntax features that have been added
in 3.5 and 3.6:

(1) Formatted string literals, such as this:

        f'Name: {name}'

I generally cleaned up string literals and binary literals in the
grammar, to be in line with the current rules in Python 3.

(2) The 'async' and 'await' keywords, which can appear in a few
places:

* 'async' can appear optionally in front of 'def', 'for', and 'with'
* 'await' can appear in front of most expressions, with a priority
  much higher than most of the other operators.

In both cases, I've added to the existing rules in Python3.g4 and
the rest of JPlag won't see these as distinct (e.g., a formatted
string literal will be seen as a string literal, 'async def' will
be seen as the beginning of a function, and so on).  I think, for
our purposes, that's probably fine, though I'm open to counterargument
on that point.