Report viewer: Allow loading results by URL
When integrating with other software ([TMS](https://gitlab.com/tms-elte)
in my case), it’s useful to be able to send users to a URL that will
automatically load the results, without having to recompile the report
viewer on the live server every time (to make use of overview.json).
This pull request introduces a new `file` query parameter, which, if
provided, makes the report viewer load the results from the given URL.
Similarly to the drag-and-drop upload, the URL can point to both ZIPs
and JSONs; since the URL may not have an appropriate extension, let the
browser figure out the type based on the HTTP headers. Browsers can
figure out the MIME type of drag-and-drop files as well (usually based
on the extension), so migrate also the drag-and-drop to MIME-based type
determination, dropping the `getFileExtension` function. Also improve
error handling: instead of just logging to the browser console, present
the error using `alert` – not a five-star UX, but still much better than
nothing.